src/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
h11==0.14.0
idna==3.10
phonenumbers==8.13.51
pydantic==2.10.3
pydantic-extra-types==2.10.1
pydantic_core==2.27.1
pymongo==4.10.1
python-dateutil==2.9.0.post0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
tinydb==4.8.2
typing_extensions==4.12.2
uvicorn==0.32.1
